[REFACTORY] routes module
