+++ api/models/series.go
@@ -1,0 +1,6 @@
+package models
+
+type SerieMetadata struct {
+	Labels      []string `json:"labels"`
+	SeriesCount int      `json:"seriesCount"`
+}
